update(create-skill-app): move more stuff into packages and tighten template
