+++ packages/commerce-server/package.json
@@ -15,6 +15,7 @@
     "test": "jest"
   },
   "dependencies": {
+    "@casl/ability": "^6.0.0",
     "@skillrecordings/honeycomb-tracer": "workspace:*",
     "@skillrecordings/database": "workspace:*",
     "date-fns": "^2.28.0",
