+++ packages/commerce-server/src/index.ts
@@ -3,3 +3,5 @@ export * from './parity-coupon'
 export * from './get-calculated-price'
 export * from './get-active-merchant-coupon'
 export * from './stripe'
+export * from './ability'
+export * from './purchase-validators'
