+++ packages/create-skill-app/templates/next/package.json
@@ -11,7 +11,6 @@
     "db:schema": "prisma migrate dev --create-only"
   },
   "dependencies": {
-    "@casl/ability": "^6.0.0",
     "@mdx-js/loader": "^2.1.2",
     "@mdx-js/mdx": "^2.1.2",
     "@mdx-js/react": "^2.1.2",
