--- packages/create-skill-app/templates/next/src/lib/context.ts
@@ -1,19 +1,0 @@
-import {mockDeep, DeepMockProxy} from 'jest-mock-extended'
-import {prisma} from '@skillrecordings/database'
-import {PrismaClient} from '@skillrecordings/database'
-
-export const defaultContext: Context = {prisma}
-
-export type Context = {
-  prisma: PrismaClient
-}
-
-export type MockContext = {
-  prisma: DeepMockProxy<PrismaClient>
-}
-
-export const createMockContext = (): MockContext => {
-  return {
-    prisma: mockDeep<PrismaClient>(),
-  }
-}
