--- packages/create-skill-app/templates/next/src/utils/stripe.ts
@@ -1,5 +1,0 @@
-import Stripe from 'stripe'
-
-export const stripe = new Stripe(process.env.STRIPE_SECRET_TOKEN, {
-  apiVersion: '2020-08-27',
-})
